后台任务

+++ ueueo-framework/ueueo-background-jobs/src/main/java/com/ueueo/backgroundjobs/BackgroundJobInfo.java
@@ -26,7 +26,7 @@ public class BackgroundJobInfo {
      * Try count of this job.
      * A job is re-tried if it fails.
      */
-    private short tryCount;
+    private int tryCount;
 
     /**
      * Creation time of this job.
